feat: VS Code dark mode, unified side panel, plus button in header
- Replace cyberpunk blue-tinted dark mode with VS Code neutral grays
  (#1e1e1e, #252526, #2d2d2d) while keeping neon accent system
- Unify AddDialog and DetailPanel into a single SidePanel split view:
  clicking a download shows details, Ctrl+N opens add form in same panel
- Add plus (+) button in top-right header corner for adding downloads
- Convert AddDialog modal to AddPanel inline component
- Refactor DetailPanel to content-only (rendered inside SidePanel wrapper)
- Add sidePanelMode store with openAddPanel/openDetailPanel/closeSidePanel
- Use lg breakpoint for desktop panel to avoid tablet squeeze
- Delete unused SVG logo files, update README to use PNG logo
- Update sidebar trigger icon from search to plus
